refactor(AutoComplete): tree shakeable
